refactor: remove TableEngine
