+++ src/common/meta/src/ddl/truncate_table.rs
@@ -27,8 +27,8 @@ use serde::{Deserialize, Serialize};
 use snafu::{ensure, ResultExt};
 use store_api::storage::RegionId;
 use strum::AsRefStr;
-use table::engine::TableReference;
 use table::metadata::{RawTableInfo, TableId};
+use table::table_reference::TableReference;
 
 use super::utils::handle_retry_error;
 use crate::ddl::utils::handle_operate_region_error;
