+++ src/table/src/lib.rs
@@ -15,13 +15,13 @@
 #![feature(assert_matches)]
 
 pub mod dist_table;
-pub mod engine;
 pub mod error;
 pub mod metadata;
 pub mod predicate;
 pub mod requests;
 pub mod stats;
 pub mod table;
+pub mod table_reference;
 pub mod test_util;
 pub mod thin_table;
